+++ forms/SpryngConfigureForm.php
@@ -2,13 +2,12 @@
 
 namespace humhub\modules\sms\forms;
 
-use humhub\modules\ui\form\widgets\ActiveField;
-use humhub\modules\ui\form\widgets\ActiveForm;
+use humhub\widgets\form\ActiveField;
+use humhub\widgets\form\ActiveForm;
 use Yii;
 
 class SpryngConfigureForm extends SmsProviderConfigureForm
 {
-
     // required
     public $username_spryng;
     public $password_spryng;
@@ -22,10 +21,10 @@ class SpryngConfigureForm extends SmsProviderConfigureForm
      */
     public function rules()
     {
-        return array_merge(parent::rules(), array(
-            array(['username_spryng', 'password_spryng'], 'required'),
-            array(['route_spryng', 'service_spryng', 'allowlong_spryng'], 'safe')
-        ));
+        return array_merge(parent::rules(), [
+            [['username_spryng', 'password_spryng'], 'required'],
+            [['route_spryng', 'service_spryng', 'allowlong_spryng'], 'safe'],
+        ]);
     }
 
     /**
@@ -35,13 +34,13 @@ class SpryngConfigureForm extends SmsProviderConfigureForm
      */
     public function attributeLabels()
     {
-        return array_merge(parent::attributeLabels(), array(
+        return array_merge(parent::attributeLabels(), [
             'username_spryng' => Yii::t('SmsModule.base', 'Username'),
             'password_spryng' => Yii::t('SmsModule.base', 'Password'),
             'route_spryng' => Yii::t('SmsModule.base', 'Select the Spryng route (default: BUSINESS)'),
             'service_spryng' => Yii::t('SmsModule.base', 'Reference tag to create a filter in statistics'),
-            'allowlong_spryng' => Yii::t('SmsModule.base', 'Allow Messages > 160 characters (default: not allowed -> currently not supported, as characters are limited by the view)')
-        ));
+            'allowlong_spryng' => Yii::t('SmsModule.base', 'Allow Messages > 160 characters (default: not allowed -> currently not supported, as characters are limited by the view)'),
+        ]);
     }
 
     /**
@@ -50,7 +49,7 @@ class SpryngConfigureForm extends SmsProviderConfigureForm
      */
     public function attributeNames()
     {
-        return array_merge(parent::attributeNames(), array('username_spryng', 'password_spryng', 'route_spryng', 'service_spryng', 'allowlong_spryng'));
+        return array_merge(parent::attributeNames(), ['username_spryng', 'password_spryng', 'route_spryng', 'service_spryng', 'allowlong_spryng']);
     }
 
     /**
@@ -58,7 +57,7 @@ class SpryngConfigureForm extends SmsProviderConfigureForm
      *
      * @param ActiveForm $activeForm
      * @param string $attributeName the attributes name
-     * @return ActiveField | \yii\bootstrap\ActiveField | null
+     * @return ActiveField | \humhub\widgets\form\ActiveField | null
      */
     public function getActiveFormElement($activeForm = null, $attributeName = null)
     {
@@ -67,17 +66,17 @@ class SpryngConfigureForm extends SmsProviderConfigureForm
         }
 
         switch ($attributeName) {
-            case 'username_spryng' :
+            case 'username_spryng':
                 return $activeForm->field($this, 'username_spryng')->textInput();
-            case 'password_spryng' :
+            case 'password_spryng':
                 return $activeForm->field($this, 'password_spryng')->passwordInput();
-            case 'route_spryng' :
+            case 'route_spryng':
                 return $activeForm->field($this, 'route_spryng')->textInput();
-            case 'service_spryng' :
+            case 'service_spryng':
                 return $activeForm->field($this, 'service_spryng')->textInput();
-            case 'allowlong_spryng' :
+            case 'allowlong_spryng':
                 return $activeForm->field($this, 'allowlong_spryng')->checkbox();
-            default :
+            default:
                 return parent::getActiveFormElement($activeForm, $attributeName);
         }
     }