+++ messages/eu/base.php
@@ -32,7 +32,7 @@ return array (
   'Save Configuration' => '',
   'Security error. Please contact an administrator to check the module configuration.' => '',
   'Select the Spryng route (default: BUSINESS)' => '',
-  'Send' => '',
+  'Send' => 'Bidali',
   'Send SMS' => '',
   'Send a SMS' => '',
   'Send a SMS to ' => '',