+++ docs/CHANGELOG.md
@@ -1,9 +1,15 @@
 Changelog
 =========
 
-0.8.0 (Unreleased)
-------------------------
+0.8.0 (August 28, 2025)
+-----------------------
 - Fix #19: Update module resources path
+- Enh #22: Fix JavaScript tag attribute "nonce"
+- Enh #23: Migration to Bootstrap 5 for HumHub 1.18
+
+0.7.7 (November 13, 2024)
+-------------------------
+- Enh #20: Use PHP CS Fixer
 
 0.7.6 (February 2, 2023)
 ------------------------